Provide default implementation of dataset_fn for ODPS data source
