+++ elasticdl/python/common/model_utils.py
@@ -1,6 +1,7 @@
 import importlib.util
 import os
 
+from elasticdl.python.common.constants import ODPSConfig
 from elasticdl.python.common.log_utils import default_logger as logger
 from elasticdl.python.worker.prediction_outputs_processor import (
     BasePredictionOutputsProcessor,
@@ -123,9 +124,22 @@ def get_model_spec(
             "inherited from BasePredictionOutputsProcessor. "
             "Prediction outputs may not be processed correctly."
         )
+
+    # If ODPS data source is used, dataset_fn is optional
+    dataset_fn_required = not all(
+        k in os.environ
+        for k in (
+            ODPSConfig.PROJECT_NAME,
+            ODPSConfig.ACCESS_ID,
+            ODPSConfig.ACCESS_KEY,
+        )
+    )
+
     return (
         model,
-        _get_spec_value(dataset_fn, model_zoo, default_module, required=True),
+        _get_spec_value(
+            dataset_fn, model_zoo, default_module, required=dataset_fn_required
+        ),
         _get_spec_value(loss, model_zoo, default_module, required=True),
         _get_spec_value(optimizer, model_zoo, default_module, required=True),
         _get_spec_value(
