+++ scripts/client_test.sh
@@ -81,7 +81,7 @@ elif [[ "$JOB_TYPE" == "odps" ]]; then
       --model_zoo=model_zoo \
       --model_def=odps_iris_dnn_model.odps_iris_dnn_model.custom_model \
       --training_data=$ODPS_TABLE_NAME \
-      --data_reader_params='columns=["sepal_length", "sepal_width", "petal_length", "petal_width", "class"]' \
+      --data_reader_params='columns=["sepal_length", "sepal_width", "petal_length", "petal_width", "class"],label_col="class"' \
       --envs="ODPS_PROJECT_NAME=$ODPS_PROJECT_NAME,ODPS_ACCESS_ID=$ODPS_ACCESS_ID,ODPS_ACCESS_KEY=$ODPS_ACCESS_KEY,ODPS_ENDPOINT=" \
       --num_epochs=2 \
       --master_resource_request="cpu=0.2,memory=1024Mi" \
